fixes workflow default rule when custimizing is missing
When no users are customized for a given cost center and role on the
customer backend, the logic for checking if the agents for the current
step also all appear in later workflow steps (to avoid duplicated
workitems/decisions) was incorrect.
When there are no agents for the current step and no users for the
forecasted steps, then by definition the agents of the current step
cannot appear in later steps, as implemented in this commit.