changes how obligatory fields are checked for clearing supplier form
The supplier form is cleared after successfully sending the supplier
creation request via e-mail, in order to avoid sending the request
multiple times. Clearing the fields, marks them as obligatory, prompting
the user to enter the data again, which is the opposite of the desired
effect. DCF fields are now no longer marked as obligatory and instead
the checks are performed in the BAdI. Missing fields are marked for the
user, but only after submitting the form. This is a drawback, compared
to the checking and prompt when using the DCF function, because the
check is not executed after the field loses focus and instead is checked
in the backend.